[Refactor] modeWrap 폴더 생성에 따른 import 경로 수정 #418
